only run service setup during first start

+++ etc/cont-init.d/10-prepare-rclone.sh
@@ -25,7 +25,7 @@ if [ ! -f "/config/rclone.conf" ]; then
 
   # create the config
   /usr/bin/rclone config create "$rclone_remote" crypt remote="$remote" password="$pwobscure" password2="$pwobscurehash"
-elif [[ -z $(cat /config/rclone.conf | grep "$rclone_remote") ]]
+elif [[ -z $(cat /config/rclone.conf | grep "$rclone_remote") ]]; then
   # remote does not exist but rclone.conf does exist, create new remote
   echo "Creating remote config $rclone_remote"
 
@@ -39,33 +39,34 @@ elif [[ -z $(cat /config/rclone.conf | grep "$rclone_remote") ]]
   /usr/bin/rclone config create "$rclone_remote" crypt remote="$remote" password="$pwobscure" password2="$pwobscurehash"
 fi
 
-# most of this only needs to be done on first container start. This should be detected by touching a file
+if [ ! -f /setupcontainer ]; then
+  # remove password from env
+  unset PASSWORD
+  unset PASSWORD2
 
-# remove password from env
-unset PASSWORD
-unset PASSWORD2
+  # create folders 
+  # the merged fs - local cache for gdrive - new local only files
+  mkdir -p /local/{cache,gdrive} /config/log
 
-# create gdrive-rclone service
-echo "Creating rclone service file"
+  # link logs for rclone
+  ln -sf /config/log /var/log/rclone
 
-cd /etc/services.d/rclone || exit 1
+  # create gdrive-rclone service
+  echo "Creating rclone service file"
+  cd /etc/services.d/rclone || exit 1
 
-sed -i "s,<cache-size>,$LOCAL_CACHE_SIZE,g" run finish
-sed -i "s,<cache-time>,$LOCAL_CACHE_TIME,g" run finish
-sed -i "s,<gdrive-rclone>,$rclone_remote,g" run finish
+  sed -i "s,<cache-size>,$LOCAL_CACHE_SIZE,g" run finish
+  sed -i "s,<cache-time>,$LOCAL_CACHE_TIME,g" run finish
+  sed -i "s,<gdrive-rclone>,$rclone_remote,g" run finish
 
-# create mountable folders 
-# the merged fs - local cache for gdrive - new local only files
-mkdir -p /remote /local/{cache,gdrive} /config/log
+  echo "Creating cron task"
+  mkdir -p /etc/crontabs
+  echo "0 */6 * * * /usr/bin/rclone move /local/gdrive $rclone_remote: --config /config/rclone.conf --log-file /var/log/rclone/upload.log --log-level INFO --delete-empty-src-dirs --fast-list --min-age 6h" > /etc/crontabs/root
+  crontab /etc/crontabs/root
 
-# logs for rclone
-ln -sf /config/log /var/log/rclone
-
-# create cronjob task
-echo "Creating cron task"
-mkdir /etc/crontabs
-echo "0 */6 * * * /usr/bin/rclone move /local/gdrive $rclone_remote: --config /config/rclone.conf --log-file /var/log/rclone/upload.log --log-level INFO --delete-empty-src-dirs --fast-list --min-age 6h" > /etc/crontabs/root
-crontab /etc/crontabs/root
+  # so we know the container has already been setup
+  touch /setupcontainer
+fi
 
 echo "Mounting mergerfs"
 /usr/bin/mergerfs /local/gdrive:/gdrive-cloud /remote -o rw,use_ino,allow_other,func.getattr=newest,category.action=all,category.create=ff,cache.files=auto-full,nonempty
